Uploading the Files
This is the initial commit which pushes the tweaked version from the generator-create-rc-admin which I made as a part of my C4GT project.
This project is the initialization for the contributions of the Google Summer of Code participants to put out their details and check if they are still
contrbuting into their project or got busy with their jobs. Anyhow, The site is not much generic yet. But I wish some day soon into this world, I can make
attempts to grow this site into some good level of project. Wishing myself a good luck with this commit message.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>