Fix: Header message and remove Auth
This patch is intended for removing the wrong heading message which was X-Admin because this site is bloated from my react-admin generator so the heading
was fixed. And also, There was Auth which was not at all required for the scope of this project. So, For now temporarily I removed the auth from Admin
Component. There is still a possibility to remove the file which provides auth but I am bit sloppy for now and won't make this done straight away. Let's
get this project into the development phase maybe sooner or later.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>